+++ client/src/App.tsx
@@ -6,6 +6,9 @@ import LandingPage from './pages/LandingPage'
 import MainPage from './pages/MainPage'
 import PortfolioPage from './pages/PortfolioPage'
 import LearningPage from './pages/LearningPage'
+import LearningCategoryListPage from './pages/LearningCategoryListPage'
+import LearningDocListPage from './pages/LearningDocListPage'
+import LearningDocPage from './pages/LearningDocPage'
 import ColumnPage from './pages/ColumnPage'
 import ProjectPage from './pages/ProjectPage'
 import AboutPage from './pages/AboutPage'
@@ -22,6 +25,18 @@ function App() {
           <Route path="about" element={<AboutPage />} />
           <Route path="portfolio" element={<PortfolioPage />} />
           <Route path="learning" element={<LearningPage />} />
+          <Route
+            path="learning/info-engineer/:categoryId/:docSlug"
+            element={<LearningDocPage />}
+          />
+          <Route
+            path="learning/info-engineer/:categoryId"
+            element={<LearningDocListPage />}
+          />
+          <Route
+            path="learning/info-engineer"
+            element={<LearningCategoryListPage />}
+          />
           <Route path="column" element={<ColumnPage />} />
           <Route path="project" element={<ProjectPage />} />
         </Route>